Add files via upload
Note: In this project, I performed data cleaning and created interactive dashboards to visualize the insights effectively. Data Cleaning: The raw data was cleaned and pre- 
          processed in Excel, ensuring accuracy and consistency.  Dashboards: I developed dashboards in both Power BI and Excel to provide dynamic visualizations of the cleaned 
          data.

Files Provided: You can find the cleaned dataset in the Excel file along with the Excel dashboard. The Power BI dashboard is provided as a separate .pbix file.